Add files via upload
An efficient algorithm to splits a given array of positive integers into two arrays of equal sum. A solution may not always exist, and the script will tell you if it doesn't. It uses recursion, and is optimized to solve it in less than O(n), where n is the length of the given array.